requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
copyright==1.0.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.2
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.3
django-browser-reload==1.7.0
django-crispy-forms==2.0
django-datatables-view==1.20.0
django-redis==5.2.0
django-rosetta==0.9.9
django-select2==8.1.1
idna==3.4
install==1.3.5
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
pandas==2.0.0
polib==1.2.0
psycopg2-binary==2.9.6
pyclean==2.7.0
pycodestyle==2.10.0
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
regex==2023.5.5
requests==2.31.0
rosetta==0.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
XlsxWriter==3.1.0
